+++ modules/SofaBoundaryCondition/PartialLinearMovementConstraint.inl
@@ -385,7 +385,7 @@ void PartialLinearMovementConstraint<DataTypes>::projectJacobianMatrix(const cor
                                         {
                                             for (auto colIt = itRow.begin(); colIt != itRow.end(); colIt++)
                                             {
-                                                if (index == colIt.index())
+                                                if (index == (unsigned int)colIt.index())
                                                 {
                                                     Deriv b = colIt.val();
                                                     for (unsigned int j = 0; j < btype.size(); j++) if (btype[j]) b[j] = 0.0;
